requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==3.2
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sqlparse==0.4.3
toml==0.10.2
django-posts~=0.3.0
python-dotenv~=1.0.1